added separate utils file for LLM purpose
