+++ README.md
@@ -11,7 +11,7 @@
 1. Check if your `Node.js` version is >= **14**.
 2. Run `npm install` to install the dependencies.
 3. replace your GEMINI api key in `.env.template` and rename the file  to `.env`
-<!-- 3. you need to either be running a local model on `http://localhost:1234` or change the openAI config in `src/utils.js` -->
+<!-- 3. you need to either be running a local model on `http://localhost:1234` or change the openAI config in `src/llm-utils.js` -->
 
 ## Developing
 
@@ -38,12 +38,12 @@ The main source files are:
 - `src/popup/popup.jsx` The popup window that shows when you press the extension icon
 - `src/background/background.js` Persists and facilitates planning
 - `src/contentScript/contentScript.js` Executes on web pages, executes actions and scrapes elements
-- `src/utils.js` Helper functions
+- `src/llm-utils.js` Helper functions
 
 ### Current Flow
 1. User clicks extension icon, enters and submits goal prompt
 2. Popup.jsx sends a message to background.js with request type 'new_goal'
-3. background.js calls promptToFirstStep() in utils.js
+3. background.js calls promptToFirstStep() in llm-utils.js
 4. LLM returns starting URL which feeds into navURL()
 5. navURL will open a new tab with the starting URL
 6. getNextStep() is called which will get contentScript to extract clickable elements and send it to LLM to generate next step
